main/requirements.txt
alembic==0.8.5
aniso8601==1.1.0
click==6.6
facepy==1.0.8
Flask==0.10.1
Flask-Cache==0.13.1
flask-marshmallow==0.6.2
Flask-Migrate==1.8.0
Flask-RESTful==0.3.5
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-Webpack==0.1.0
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.4
MarkupSafe==0.23
marshmallow==2.7.0
marshmallow-jsonapi==0.7.0
marshmallow-sqlalchemy==0.8.1
numpy==1.11.0
pandas==0.18.0
PyMySQL==0.7.2
python-dateutil==2.5.2
python-editor==1.0
pytz==2016.3
redis==2.10.5
requests==2.9.1
rq==0.5.6
six==1.10.0
SQLAlchemy==1.0.12
Werkzeug==0.11.5
